make helper format description
